backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.311
langchain-community==0.0.11
openai==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.14.0
chromadb==0.4.13
pydantic==2.4.2
python-dotenv==1.0.0
websockets==12.0
jinja2==3.1.2
pandoc==1.0.0
pytest==7.4.3
httpx==0.25.2
typing-inspect==0.9.0
typing-extensions==4.8.0
tiktoken==0.5.1 